pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
poetry = "^1.8.2"
streamlit = "1.33.0"
streamlit-chat = "^0.1.1"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
langchain-community = "^0.0.32"
sentence-transformers = "^2.6.1"
openai = "^1.17.1"
pypdf = "^4.2.0"
chromadb = "^0.4.24"
tiktoken = "^0.6.0"
pillow = "^10.3.0"
networkx = "^3.3"
streamlit-shadcn-ui = "^0.1.18"
llama-index = "0.10.28"
llama-index-embeddings-openai = "0.1.7"
llama-index-vector-stores-pinecone = "^0.1.4"
pysqlite3 = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
